% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregorian.R
\name{as_gregorian}
\alias{as_gregorian}
\title{Convert to a Gregorian date}
\usage{
as_gregorian(date)
}
\arguments{
\item{date}{A gregorian vector object}
}
\description{
Convert to a Gregorian date
}
\examples{
as_gregorian("2016-01-01")
as_gregorian(Sys.Date())
tibble::tibble(
  x = seq(as.Date("2025-01-01"), as.Date("2025-12-31"), by = "day"),
  y = as_gregorian(x)
)
}
