% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hebrew.R
\name{as_hebrew}
\alias{as_hebrew}
\title{Convert to a Hebrew date}
\usage{
as_hebrew(date)
}
\arguments{
\item{date}{Vector of dates on some calendar}
}
\description{
Convert to a Hebrew date
}
\examples{
as_hebrew("2016-01-01")
as_hebrew(Sys.Date())
tibble::tibble(
  x = seq(as.Date("2025-01-01"), as.Date("2025-12-31"), by = "day"),
  y = as_hebrew(x)
)
}
