% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islamic.R
\name{as_islamic}
\alias{as_islamic}
\title{Convert to an Islamic date}
\usage{
as_islamic(date)
}
\arguments{
\item{date}{Vector of dates on some calendar}
}
\description{
Convert to an Islamic date
}
\examples{
as_islamic("2016-01-01")
as_islamic(Sys.Date())
tibble::tibble(
  x = seq(as.Date("2025-01-01"), as.Date("2025-12-31"), by = "day"),
  y = as_islamic(x)
)
}
