% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.R
\name{as_iso}
\alias{as_iso}
\title{Convert to a ISO date}
\usage{
as_iso(date)
}
\arguments{
\item{date}{Vector of dates on some calendar}
}
\description{
Convert to a ISO date
}
\examples{
as_iso("2016-01-01")
as_iso(Sys.Date())
tibble::tibble(
  x = seq(as.Date("2025-01-01"), as.Date("2025-12-31"), by = "day"),
  y = as_iso(x)
)
}
