% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian.R
\name{as_julian}
\alias{as_julian}
\title{Convert to julian_date}
\usage{
as_julian(date, ...)
}
\arguments{
\item{date}{Date on some calendar}

\item{...}{Additional arguments}
}
\value{
An julian_date object
}
\description{
Convert to julian_date
}
\examples{
as_julian("2016-01-01")
as_julian(Sys.Date())
}
