% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed.R
\name{as_rd}
\alias{as_rd}
\title{Convert date to rd_fixed date}
\usage{
as_rd(date, ...)
}
\arguments{
\item{date}{Date on some calendar}

\item{...}{Additional arguments}
}
\value{
An rd_fixed object representing RD fixed date
}
\description{
Convert date to rd_fixed date
}
\examples{
as_rd("2016-01-01")
as_rd(Sys.Date())
}
