% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roman.R
\name{as_roman}
\alias{as_roman}
\title{Convert to a Roman date}
\usage{
as_roman(date)
}
\arguments{
\item{date}{Vector of dates on some calendar}
}
\description{
Convert to a Roman date
}
\examples{
as_roman("2016-01-01")
as_roman(Sys.Date())
tibble::tibble(
  x = seq(as.Date("2025-01-01"), as.Date("2025-12-31"), by = "day"),
  y = as_roman(x)
)
}
