% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babylonian.R
\name{babylonian_date}
\alias{babylonian_date}
\alias{as_babylonian}
\title{Babylonian dates}
\usage{
babylonian_date(year, month, leap, day)

as_babylonian(date)
}
\arguments{
\item{year}{Numeric vector of years}

\item{month}{Numeric vector of months}

\item{leap}{Logical vector of leap months}

\item{day}{Numeric vector of days}
}
\value{
A babylonian vector object
}
\description{
Babylonian dates
}
\examples{
babylonian_date(2025, 6, FALSE, 1:10)
}
\seealso{
\link{cal_babylonian}
}
