% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bahai.R
\name{bahai_date}
\alias{bahai_date}
\alias{as_bahai}
\title{Bahai dates}
\usage{
bahai_date(major, cycle, year, month, day)

as_bahai(date)
}
\arguments{
\item{major}{A numeric vector of major periods}

\item{cycle}{A numeric vector of cycles}

\item{year}{A numeric vector of years within the cycles}

\item{month}{A numeric vector of months}

\item{day}{A numeric vector of days}

\item{date}{A numeric vector of dates}
}
\value{
A bahai vector object
}
\description{
The Bahai calendar is a solar calendar comprising 18 months, with four or five
intercalary days each year. The New Year is at the northern Spring equinox,
corresponding to 21 March on the Gregorian calendar.
}
