% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bahai.R
\name{bahai_new_year}
\alias{bahai_new_year}
\alias{naw_ruz}
\alias{feast_of_ridvan}
\alias{birth_of_the_bab}
\title{Bahai holidays}
\usage{
bahai_new_year(year)

naw_ruz(year)

feast_of_ridvan(year)

birth_of_the_bab(year)
}
\arguments{
\item{year}{The year on the Gregorian calendar}
}
\value{
A vector of dates on the Gregorian calendar
}
\description{
Dates are returned as Gregorian dates
}
\examples{
tibble::tibble(
  year = 2025:2030,
  new_year = bahai_new_year(year),
  naw_ruz =naw_ruz(year),
  ridvan = feast_of_ridvan(year),
  birth_bab = birth_of_the_bab(year)
)
}
