% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bali.R
\name{balinese_date}
\alias{balinese_date}
\alias{as_balinese}
\title{Balinese Pawukon dates}
\usage{
balinese_date(
  luang,
  dwiwara,
  triwara,
  caturwara,
  pancawara,
  sadwara,
  saptawara,
  asatawara,
  sangawara,
  daswara
)

as_balinese(date)
}
\arguments{
\item{luang}{A numeric vector}

\item{dwiwara}{A numeric vector}

\item{triwara}{A numeric vector}

\item{caturwara}{A numeric vector}

\item{pancawara}{A numeric vector}

\item{sadwara}{A numeric vector}

\item{saptawara}{A numeric vector}

\item{asatawara}{A numeric vector}

\item{sangawara}{A numeric vector}

\item{daswara}{A numeric vector}

\item{date}{A vector of dates on some calendar.}
}
\description{
Balinese Pawukon dates
}
