% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hebrew.R
\docType{data}
\name{cal_hebrew}
\alias{cal_hebrew}
\title{Hebrew calendar dates}
\format{
An object of class \code{calcal} of length 8.
}
\usage{
cal_hebrew
}
\description{
Create a Hebrew date object.
}
\examples{
heb <- new_date(year = 5785, month = 3, day = 2:4, calendar = cal_hebrew)
heb
as_date(heb, calendar = cal_gregorian)
as_date(Sys.Date(), calendar = cal_hebrew)
tibble::tibble(
  x = seq(as.Date("2025-01-01"), as.Date("2025-12-31"), by = "day"),
  y = as_date(x, calendar = cal_hebrew),
  z = as_date(x, calendar = cal_gregorian)
)
}
\keyword{datasets}
