% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islamic.R
\docType{data}
\name{cal_islamic}
\alias{cal_islamic}
\title{Work with Islamic dates}
\format{
An object of class \code{calcal} of length 8.
}
\usage{
cal_islamic
}
\description{
Work with Islamic dates
}
\examples{
as_date("2016-01-01", calendar = cal_islamic)
as_date(Sys.Date(), calendar = cal_islamic)
tibble::tibble(
  x = new_date(year = 2025, month = 5, day = 1:31, calendar = cal_gregorian),
  y = as_date(x, calendar = cal_islamic)
)
new_date(year = 2025, month = 4, day = 19:30, calendar = cal_islamic)
}
\keyword{datasets}
