% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.R
\docType{data}
\name{cal_iso}
\alias{cal_iso}
\title{ISO calendar dates}
\format{
An object of class \code{calcal} of length 8.
}
\usage{
cal_iso
}
\description{
Work with ISO 8601 date objects. Weeks are defined as starting on Mondays. Week 1
is the first week with at least 4 days in the year. Equivalently, it is the week
containing 4 January. There is no week 0; instead week 1 of a year may begin in
the previous calendar year.
}
\details{
More flexible week numbering is possible using Gregorian dates with \code{\link{week_of_year}()}.
}
\examples{
iso <- new_date(year = 2025, week = 23, day = 2:4, calendar = cal_iso)
iso
as_date(iso, calendar = cal_gregorian)
greg <- new_date(year = 2025, month = 1, day = 1:4, calendar = cal_gregorian)
as_date(greg, calendar = cal_iso)
}
\seealso{
\code{\link{week_of_year}()}
}
\keyword{datasets}
