% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian.R
\docType{data}
\name{cal_julian}
\alias{cal_julian}
\title{Work with Julian dates}
\format{
An object of class \code{calcal} of length 8.
}
\usage{
cal_julian
}
\description{
Work with Julian dates
}
\examples{
as_date("2016-01-01", calendar = cal_julian)
as_date(Sys.Date(), calendar = cal_julian)
tibble::tibble(
  x = seq(as.Date("2025-01-01"), as.Date("2025-12-31"), by = "day"),
  y = as_date(x, calendar = cal_gregorian),
  z = as_date(x, calendar = cal_julian)
)
new_date(year = 2025, month = 4, day = 19:30, calendar = cal_julian)
}
\keyword{datasets}
