% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roman.R
\docType{data}
\name{cal_roman}
\alias{cal_roman}
\title{Work with Roman calendar dates}
\format{
An object of class \code{calcal} of length 8.
}
\usage{
cal_roman
}
\description{
Create a Roman date object.
}
\details{
roman_date(66, 4, 1, 1, FALSE)
as_roman(Sys.Date())
}
\keyword{datasets}
