% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcal-package.R
\docType{package}
\name{calcal-package}
\alias{calcal}
\alias{calcal-package}
\title{calcal: Calendrical Calculations}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An R implementation of the algorithms described in Reingold and Dershowitz (4th ed., Cambridge University Press, 2018) \doi{10.1017/9781107415058}, allowing conversion between many different calendar systems. Cultural and religious holidays from several calendars can be calculated.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkg.robjhyndman.com/calcal/}
  \item \url{https://github.com/robjhyndman/calcal}
  \item Report bugs at \url{https://github.com/robjhyndman/calcal/issues}
}

}
\author{
\strong{Maintainer}: Rob Hyndman \email{Rob.Hyndman@monash.edu} (\href{https://orcid.org/0000-0002-2140-5352}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Edward Reingold (Original Lisp code) [copyright holder, contributor]
  \item Nachum Dershowitz (Original Lisp code) [copyright holder, contributor]
}

}
\keyword{internal}
