% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinese.R
\name{chinese_new_year}
\alias{chinese_new_year}
\alias{qing_ming}
\title{Chinese holidays}
\usage{
chinese_new_year(year)

qing_ming(year)
}
\arguments{
\item{year}{The year on the Gregorian calendar}
}
\description{
Dates are returned as Gregorian dates
}
