% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregorian.R
\name{day_number}
\alias{day_number}
\alias{days_remaining}
\title{Day of year}
\usage{
day_number(g_date)

days_remaining(g_date)
}
\arguments{
\item{g_date}{A Gregorian date object}
}
\value{
A numeric vector of the same length as g_date
}
\description{
Day number in year or days remaining in year given a Gregorian date
}
\examples{
day_number(gregorian(2025, 5, 2))
days_remaining(gregorian(2025, 5, 2))

}
