% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granularities.R
\name{granularities}
\alias{granularities}
\alias{granularity}
\title{Canonical granularities for a calendar}
\usage{
granularities(dates)

granularity(dates, granularity)
}
\arguments{
\item{dates}{A vector of calendar dates}

\item{granularity}{A character string indicating the granularity to extract}
}
\description{
Canonical granularities for a calendar

Extract a granularity from a calendar
}
\examples{
granularities(iso_date(2025, 23, 2:4))
granularities(gregorian_date(2025, 4, 19:30))
granularities(as_roman(Sys.Date()))
granularity(iso_date(2025, 23, 2:4), "week")
granularity(gregorian_date(2025, 4, 19:21), "month")
granularity(as_roman(Sys.Date()), "event")
}
