% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregorian.R
\name{day_of_week}
\alias{day_of_week}
\alias{day_of_year}
\alias{days_remaining}
\alias{day_of_month}
\alias{month_of_year}
\alias{week_of_year}
\title{Extract parts of a Gregorian date}
\usage{
day_of_week(date, numeric = FALSE, first_day = "Monday", abbreviate = FALSE)

day_of_year(date)

days_remaining(date)

day_of_month(date)

month_of_year(date)

week_of_year(date, first_day = "Monday")
}
\arguments{
\item{date}{A vector of Gregorian dates}

\item{numeric}{Logical. Return a numeric vector if TRUE with 1 denoting the \code{first_day}}

\item{first_day}{Character denoting first day of the week. Ignored if \code{numeric} is \code{FALSE}.
Default is \code{"Monday"}}

\item{abbreviate}{Logical. Return abbreviated day names if \code{TRUE}. Ignored if \code{numeric} is \code{TRUE}.}
}
\description{
Extract days, weeks, or months from a vector of Gregorian dates.
}
\details{
\code{week_of_year()} returns the ISO 8601 week number with \code{first_day} as Monday.
Under this standard, week 1 of a year is defined as the first week with at least 4 days in the year;
equivalently, it is the week  containing 4 January.

\code{days_remaining()} returns the number of days remaining in the year. Other functions should be
self-explanatory.
}
\examples{
april2025 <- gregorian(2025, 4, 1:30)
day_of_week(april2025)
day_of_month(april2025)
day_of_year(april2025)
days_remaining(april2025)
month_of_year(april2025)
week_of_year(april2025)
}
