% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregorian.R
\name{gregorian}
\alias{gregorian}
\alias{as_gregorian}
\title{Gregorian dates}
\usage{
gregorian(year = integer(), month = integer(), day = integer())

as_gregorian(date, ...)
}
\arguments{
\item{year}{A numeric vector of years}

\item{month}{A numeric vector of months}

\item{day}{A numeric vector of days}

\item{date}{Vector of dates on some calendar}

\item{...}{Additional arguments not currently used}
}
\value{
A gregorian vector object
}
\description{
Create a Gregorian date object. Dates before the establishment of the Gregorian calendar are computed retrospectively.
}
\examples{
gregorian(2025, 4, 19)
as_gregorian("2016-01-01")
as_gregorian(Sys.Date())
tibble::tibble(
  x = seq(as.Date("2025-01-01"), as.Date("2025-12-31"), by = "day"),
  y = as_gregorian(x)
)
}
