% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregorian.R
\docType{data}
\name{gregorian_date}
\alias{gregorian_date}
\alias{cal_gregorian}
\title{Gregorian calendar dates}
\format{
An object of class \code{calcal} of length 8.
}
\usage{
gregorian_date(year = integer(), month = integer(), day = integer())

cal_gregorian
}
\arguments{
\item{year}{A numeric vector of years}

\item{month}{A numeric vector of months}

\item{day}{A numeric vector of days}
}
\value{
A gregorian vector object
}
\description{
Create a Gregorian date object.
}
\examples{
gregorian_date(2025, 4, 19:30)
new_date(year = 2025, month = 3, day = 2:4, calendar = cal_gregorian)
as_date(Sys.Date(), calendar = cal_gregorian)
tibble::tibble(
  x = seq(as.Date("2025-01-01"), as.Date("2025-12-31"), by = "day"),
  z = as_date(x, calendar = cal_gregorian)
)
}
\keyword{datasets}
