% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hebrew.R
\docType{data}
\name{cal_hebrew}
\alias{cal_hebrew}
\alias{hebrew_date}
\alias{as_hebrew}
\title{Work with Hebrew calendar dates}
\format{
An object of class \code{calcal} of length 8.
}
\usage{
cal_hebrew

hebrew_date(year, month, day)

as_hebrew(date)
}
\arguments{
\item{year}{A numeric vector of years}

\item{month}{A numeric vector of months}

\item{day}{A numeric vector of days}

\item{date}{Vector of dates on some calendar}
}
\value{
A hebrew vector object
}
\description{
Work with Hebrew calendar dates
}
\examples{
heb <- new_date(year = 5785, month = 3, day = 2:4, calendar = cal_hebrew)
heb
as_date(heb, calendar = cal_gregorian)
as_date(Sys.Date(), calendar = cal_hebrew)
tibble::tibble(
  x = seq(as.Date("2025-01-01"), as.Date("2025-12-31"), by = "day"),
  y = as_date(x, calendar = cal_hebrew),
  z = as_date(x, calendar = cal_gregorian)
)
hebrew_date(5785, 3, 2:4)
as_gregorian(hebrew_date(5785, 3, 2:4))
as_hebrew(gregorian_date(2025, 1, 1:31))
as_hebrew("2016-01-01")
as_hebrew(Sys.Date())
tibble::tibble(
  x = seq(as.Date("2025-01-01"), as.Date("2025-12-31"), by = "day"),
  y = as_hebrew(x)
)
}
\keyword{datasets}
