% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hebrew.R
\name{hebrew_date}
\alias{hebrew_date}
\title{Hebrew calendar dates}
\usage{
hebrew_date(year = integer(), month = integer(), day = integer())
}
\arguments{
\item{year}{A numeric vector of years}

\item{month}{A numeric vector of months}

\item{day}{A numeric vector of days}
}
\value{
A hebrew vector object
}
\description{
Create a Hebrew date object.
}
\examples{
hebrew_date(5785, 3, 2:4)
as_gregorian(hebrew_date(5785, 3, 2:4))
as_hebrew(gregorian_date(2025, 1, 1:31))
}
