% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icelandic.R
\name{icelandic_date}
\alias{icelandic_date}
\alias{as_icelandic}
\title{Icelandic dates}
\usage{
icelandic_date(year, season, week, weekday)

as_icelandic(date)
}
\arguments{
\item{year}{A numeric vector of years}

\item{season}{A numeric vector of seasons (1 = Summer, 2 = Winter)}

\item{week}{A numeric vector of weeks within the season}

\item{weekday}{A number vector containing day of week}

\item{date}{A numeric vector of dates}
}
\description{
Icelandic dates
}
