% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islamic.R
\name{islamic_new_year}
\alias{islamic_new_year}
\alias{mawlid}
\alias{ramadan}
\alias{eid_al_fitr}
\alias{eid_al_adha}
\title{Islamic holidays}
\usage{
islamic_new_year(year)

mawlid(year)

ramadan(year)

eid_al_fitr(year)

eid_al_adha(year)
}
\arguments{
\item{year}{A numeric vector of Gregorian years}
}
\value{
A list of dates on the Gregorian calendar
}
\description{
Functions to return Gregorian dates for various Islamic holidays
}
\examples{
tibble::tibble(
  year = 2025:2029,
  islamic_new_year = islamic_new_year(year),
  mawlid = mawlid(year),
  ramadan = ramadan(year),
  eid_al_fitr = eid_al_fitr(year),
  eid_al_adha = eid_al_adha(year)
)
ramadan(2030)
}
