% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islamic.R
\name{islamic_date}
\alias{islamic_date}
\title{Islamic calendar dates}
\usage{
islamic_date(year = integer(), month = integer(), day = integer())
}
\arguments{
\item{year}{A numeric vector of years}

\item{month}{A numeric vector of months}

\item{day}{A numeric vector of days}
}
\value{
An islamic vector object
}
\description{
Create an Islamic date object.
}
\examples{
islamic_date(2025, 4, 19:30)
}
