% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian.R
\name{julian}
\alias{julian}
\alias{as_julian}
\title{Julian dates}
\usage{
julian(year = integer(), month = integer(), day = integer())

as_julian(date, ...)
}
\arguments{
\item{year}{A numeric vector of years}

\item{month}{A numeric vector of months}

\item{day}{A numeric vector of days}

\item{date}{Vector of dates on some calendar}

\item{...}{Additional arguments not currently used}
}
\value{
A julian vector object
}
\description{
Create a Julian date object.
}
\examples{
julian(2025, 4, 19)
as_julian("2016-01-01")
as_julian(Sys.Date())
tibble::tibble(
  x = seq(as.Date("2025-01-01"), as.Date("2025-12-31"), by = "day"),
  y = as_gregorian(x),
  z = as_julian(x)
)
}
