% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian.R
\name{julian_date}
\alias{julian_date}
\title{Julian calendar dates}
\usage{
julian_date(year = integer(), month = integer(), day = integer())
}
\arguments{
\item{year}{A numeric vector of years}

\item{month}{A numeric vector of months}

\item{day}{A numeric vector of days}
}
\value{
A julian vector object
}
\description{
Create a Julian date object.
}
\examples{
julian_date(2025, 4, 19:30)
}
