% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bali.R
\name{kajeng_keliwon}
\alias{kajeng_keliwon}
\alias{tumpek}
\title{Balinese special days}
\usage{
kajeng_keliwon(year)

tumpek(year)
}
\arguments{
\item{year}{A numeric vector of Gregorian years}
}
\value{
A vector of dates on the Gregorian calendar
}
\description{
Find all occurrences of Kajeng Keliwon and Tumpek in a vector of
Gregorian years.
}
\examples{
kajeng_keliwon(2025)
tumpek(2025)

}
\seealso{
\link{balinese_date}
}
