% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_granularities.R
\name{linear_granularity}
\alias{linear_granularity}
\title{Linear Granularity}
\usage{
linear_granularity(gran_fun, gran_levels, gran_name)
}
\arguments{
\item{gran_fun}{A function that extracts a cyclic granularity from a date
variable. This function must return an integer.}

\item{gran_levels}{A character vector of the granularity levels.}

\item{gran_name}{A character string giving the name of the granularity}
}
\value{
A function that creates a linear granularity object.
}
\description{
Linear granularities are produced by combining the year and a cyclic granularity.
}
\examples{
# Compute quarter from a Gregorian date
quarter <- function(date) {
  ceiling(month_of_year(date) / 3)
}
year_quarter <- linear_granularity(quarter, paste0("Q", 1:4), "year-quarter")
year_quarter(gregorian_date(2025, 6, 25) + 1:10)
}
