% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islamic.R
\name{mawlid}
\alias{mawlid}
\title{Islamic holidays}
\usage{
mawlid(year)
}
\arguments{
\item{year}{A numeric vector of Gregorian years}
}
\value{
A list of dates on the Gregorian calendar
}
\description{
Islamic holidays
}
