% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_cal_objects.R
\name{as_date}
\alias{as_date}
\alias{new_date}
\title{Convert a date to a new calendar}
\usage{
as_date(date, calendar)

new_date(..., calendar)
}
\arguments{
\item{date}{Date vector on some calendar}

\item{calendar}{Target calendar of class "calcalcal"}

\item{...}{Named arguments denoting the granularities required for \code{calendar}.}
}
\description{
Convert a date to a new calendar

Create a new data object on a specified calendar
}
\examples{
april25 <- new_date(year = 2025, month = 4, day = 1:30, calendar = cal_gregorian)
as_date(april25, calendar = cal_iso)
}
