% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregorian.R
\name{new_gregorian_date}
\alias{new_gregorian_date}
\title{Create a new gregorian_date object}
\usage{
new_gregorian_date(year, month, day)
}
\arguments{
\item{year}{The numeric year}

\item{month}{The numeric month of year}

\item{day}{The numeric day of month}
}
\value{
A gregorian_date object
}
\description{
Create a new gregorian_date object
}
\examples{
new_gregorian_date(2025, 4, 19)
}
