% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian.R
\name{new_julian_date}
\alias{new_julian_date}
\title{Create a new julian_date object}
\usage{
new_julian_date(year, month, day)
}
\arguments{
\item{year}{The numeric year}

\item{month}{The numeric month of year}

\item{day}{The numeric day of month}
}
\value{
A julian_date object
}
\description{
Create a new julian_date object
}
\examples{
new_julian_date(2025, 4, 19)
}
