% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarlunar.R
\name{new_moons}
\alias{new_moons}
\alias{full_moons}
\title{Full moons and new moons in Gregorian years}
\usage{
new_moons(year)

full_moons(year)
}
\arguments{
\item{year}{A vector of Gregorian years}
}
\value{
A vector of dates
}
\description{
Calculate all the near-full or near-new moons in a vector of Gregorian years
}
\examples{
full_moons(2025)
new_moons(2025)
}
