% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed.R
\name{new_rd_fixed}
\alias{new_rd_fixed}
\title{Create a new rd_fixed object}
\usage{
new_rd_fixed(date)
}
\arguments{
\item{date}{The numeric date representing the RD fixed date}
}
\value{
An rd_fixed object
}
\description{
Create a new rd_fixed object
}
