% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_hindu.R
\name{old_hindu_solar_date}
\alias{old_hindu_solar_date}
\alias{old_hindu_lunar_date}
\alias{as_old_hindu_solar}
\alias{as_old_hindu_lunar}
\title{Old Hindu solar and lunar dates}
\usage{
old_hindu_solar_date(year = integer(), month = integer(), day = integer())

old_hindu_lunar_date(
  year = integer(),
  month = integer(),
  leap_month = logical(),
  day = integer()
)

as_old_hindu_solar(date)

as_old_hindu_lunar(date)
}
\arguments{
\item{year}{A numeric vector of years}

\item{month}{A numeric vector of months}

\item{day}{A numeric vector of days}

\item{leap_month}{A logical vector indicating if year is a leap year}

\item{date}{A date vector on some calendar}
}
\description{
Old Hindu solar and lunar dates
}
\examples{
gregorian_date(2025, 1, 1:31) |>
  as_old_hindu_solar()
gregorian_date(2025, 1, 1:31) |>
  as_old_hindu_lunar()
}
