% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_fixed.R
\name{rd_fixed}
\alias{rd_fixed}
\alias{as_rd}
\title{RD fixed dates}
\usage{
rd_fixed(rd = double())

as_rd(date, ...)
}
\arguments{
\item{rd}{A numeric vector representing the number of days since (and including) 01-01-01.}

\item{date}{Vector of dates on some calendar}

\item{...}{Additional arguments not currently used}
}
\value{
An rd_fixed vector object
}
\description{
Create an rd_fixed object representing an RD (Rata Die) fixed date with day 1 being 01-01-01
}
\examples{
rd_fixed(1:100)
as_rd("2016-01-01")
as_rd(Sys.Date())
}
