% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roman.R
\name{roman_date}
\alias{roman_date}
\title{Roman calendar dates}
\usage{
roman_date(
  year = integer(),
  month = integer(),
  event = integer(),
  count = integer(),
  leap = logical()
)
}
\arguments{
\item{year}{A numeric vector of years}

\item{month}{A numeric vector of months}

\item{event}{A numeric vector of events: 1 = KALENDS, 2 = NONES, 3 = IDES}

\item{count}{A numeric vector of counts:}

\item{leap}{A logical vector indicating if year is a leap year}
}
\value{
A roman vector object
}
\description{
Create a Roman date object.
}
\examples{
roman_date(66, 4, 1, 1, FALSE)
as_roman(Sys.Date())

}
