% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibetan.R
\name{tibetan_date}
\alias{tibetan_date}
\alias{as_tibetan}
\title{Tibetan calendar dates}
\usage{
tibetan_date(
  year = integer(),
  month = integer(),
  leap_month = logical(),
  day = integer(),
  leap_day = logical()
)

as_tibetan(date)
}
\arguments{
\item{year}{A numeric vector of years}

\item{month}{A numeric vector of months}

\item{leap_month}{A logical vector of leap months}

\item{day}{A numeric vector of days}

\item{leap_day}{A logical vector of leap days}

\item{date}{A vector of dates on some calendar}
}
\value{
A \code{tibetan_date} object
}
\description{
There are several Tibetan calendars. These functions implement the
official Phuglugs version of the Kalachakra calendar, which is similar
to the Hindu lunisolar calendars.
}
\examples{
gregorian_date(2025, 6, 1:10) |> as_tibetan()
}
\seealso{
\link{tibetan_new_year}
}
