% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibetan.R
\name{tibetan_new_year}
\alias{tibetan_new_year}
\alias{losar}
\title{Tibetan holidays}
\usage{
tibetan_new_year(year)

losar(t_year)
}
\arguments{
\item{year}{A vector of Gregorian years}

\item{t_year}{A vector of Tibetan years}
}
\description{
The Tibetan New Year occurs on the first day of the Tibetan calendar. These
functions calculate the date given either a Gregorian year or a Tibetan year.
Both return a Gregorian date.
}
\examples{
tibetan_new_year(2025:2028)
losar(2152:2154)
}
\seealso{
\link{tibetan_date}
}
