% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{time}
\alias{time}
\alias{as_time}
\title{Time of day}
\usage{
time(hour = integer(), minute = integer(), second = numeric())

as_time(x, ...)
}
\arguments{
\item{hour}{A numeric vector of hours}

\item{minute}{A numeric vector of minutes}

\item{second}{A numeric vector of seconds}

\item{x}{Vector of times}

\item{...}{Additional arguments not currently used}
}
\value{
A time vector object
}
\description{
Create a time object
}
\examples{
as_time(Sys.time())
}
