% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{time_of_day}
\alias{time_of_day}
\title{Time of day}
\usage{
time_of_day(hour = integer(), minute = integer(), second = numeric())
}
\arguments{
\item{hour}{A numeric vector of hours}

\item{minute}{A numeric vector of minutes}

\item{second}{A numeric vector of seconds}
}
\value{
A time_of_day vector object, stored as a vctrs record containing hours, minutes and seconds.
}
\description{
Create a time object
}
