% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_holidays.R
\name{us_memorial_day}
\alias{us_memorial_day}
\alias{us_independence_day}
\alias{us_labor_day}
\alias{us_election_day}
\alias{us_daylight_saving_start}
\alias{us_daylight_saving_end}
\alias{unlucky_fridays}
\title{US Holidays}
\usage{
us_memorial_day(year)

us_independence_day(year)

us_labor_day(year)

us_election_day(year)

us_daylight_saving_start(year)

us_daylight_saving_end(year)

unlucky_fridays(year)
}
\arguments{
\item{year}{Gregorian year}
}
\description{
Functions to return Gregorian dates for US holidays and other special days
}
\examples{
us_memorial_day(2025)
us_independence_day(2025)
us_labor_day(2025)
us_election_day(2025)
us_daylight_saving_start(2025)
us_daylight_saving_end(2025)
unlucky_fridays(2025)

}
