% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregorian.R
\name{as_gregorian}
\alias{as_gregorian}
\title{Convert to gregorian_date}
\usage{
as_gregorian(date, ...)
}
\arguments{
\item{date}{Date on some calendar}

\item{...}{Additional arguments}
}
\value{
An gregorian_date object
}
\description{
Convert to gregorian_date
}
\examples{
as_gregorian("2016-01-01")
as_gregorian(Sys.Date())
}
