% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian.R
\name{as_julian}
\alias{as_julian}
\title{Convert to a Julian date}
\usage{
as_julian(date)
}
\arguments{
\item{date}{Vector of dates on some calendar}
}
\description{
Convert to a Julian date
}
\examples{
as_julian("2016-01-01")
as_julian(Sys.Date())
tibble::tibble(
  x = seq(as.Date("2025-01-01"), as.Date("2025-12-31"), by = "day"),
  y = as_julian(x)
)
}
