% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babylonian.R
\name{babylonian_date}
\alias{babylonian_date}
\alias{as_babylonian}
\title{Babylonian dates}
\usage{
babylonian_date(year, month, leap_month, day)

as_babylonian(date)
}
\arguments{
\item{year}{Numeric vector of years}

\item{month}{Numeric vector of months}

\item{leap_month}{Logical vector of leap months}

\item{day}{Numeric vector of days}

\item{date}{Vector of dates on some calendar.}
}
\value{
A babylonian vector object
}
\description{
The classical Babylonian calendar was a lunisolar calendar with a fixed 19-year Metonic cycle.
}
\examples{
tibble::tibble(
  gregorian = gregorian_date(2025, 1, 1:31),
  babylonian = as_babylonian(gregorian)
)
babylonian_date(2025, 6, FALSE, 1:10)
}
\seealso{
\link{cal_babylonian}
}
