% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_cal_objects.R
\name{cal_calendar}
\alias{cal_calendar}
\title{Define calendar objects}
\usage{
cal_calendar(
  name,
  short_name,
  epoch,
  granularities,
  check_granularities,
  format,
  from_rd,
  to_rd
)
}
\arguments{
\item{name}{Name of calendar}

\item{short_name}{Short name of calendar}

\item{epoch}{Epoch of calendar}

\item{granularities}{Granularities of calendar}

\item{check_granularities}{Function to check granularities}

\item{format}{Functon to specify date format}

\item{from_rd}{Function to convert from RD to calendar date}

\item{to_rd}{Function to convert from calendar date to RD}
}
\description{
Generate a calendar object of class "calcalcal"
}
