% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_calendar.R, R/bahai.R, R/bali.R,
%   R/chinese.R, R/coptic.R, R/gregorian.R, R/hebrew.R, R/icelandic.R,
%   R/islamic.R, R/iso.R, R/julian.R, R/old_hindu.R, R/roman.R
\docType{data}
\name{cal_calendar}
\alias{cal_calendar}
\alias{cal_bahai}
\alias{cal_balinese}
\alias{cal_chinese}
\alias{cal_coptic}
\alias{cal_ethiopic}
\alias{cal_gregorian}
\alias{cal_hebrew}
\alias{cal_icelandic}
\alias{cal_islamic}
\alias{cal_iso}
\alias{cal_julian}
\alias{cal_old_hindu_solar}
\alias{cal_old_hindu_lunar}
\alias{cal_roman}
\title{Define calendar objects}
\format{
An object of class \code{calcal}
}
\usage{
cal_calendar(
  name,
  short_name,
  granularities,
  validate_granularities,
  format,
  from_rd,
  to_rd
)

cal_bahai

cal_balinese

cal_chinese

cal_coptic

cal_ethiopic

cal_gregorian

cal_hebrew

cal_icelandic

cal_islamic

cal_iso

cal_julian

cal_old_hindu_solar

cal_old_hindu_lunar

cal_roman
}
\arguments{
\item{name}{Name of calendar}

\item{short_name}{Short name of calendar}

\item{granularities}{Character vector with names of granularities of calendar
(e.g., for the Gregorian calendar, the granularities are \code{year}, \code{month}, and \code{day}).}

\item{validate_granularities}{Function to check granularities are valid (e.g.,
Gregorian months should be between 1 and 12).}

\item{format}{Functon to specify date format as a character string.}

\item{from_rd}{Function to convert from RD to calendar date.}

\item{to_rd}{Function to convert from calendar date to RD.}
}
\description{
Generate a calendar object of class "calcal". Examples of calendars
produced in this way include \code{cal_chinese}, \code{cal_gregorian}, \code{cal_hebrew}, \code{cal_islamic},
and \code{cal_iso}.
}
\examples{
new_date(year = 2025, month = 4, day = 19:30, calendar = cal_gregorian)
as_date("2016-01-01", calendar = cal_islamic)
as_date(Sys.Date(), calendar = cal_islamic)
tibble::tibble(
  x = new_date(year = 2025, month = 5, day = 1:31, calendar = cal_gregorian),
  y = as_date(x, calendar = cal_islamic)
)

}
\keyword{datasets}
