% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcal-package.R
\docType{package}
\name{calcal-package}
\alias{calcal}
\alias{calcal-package}
\title{calcal: Calendrical Calculations}
\description{
An R implementation of the algorithms described in Reingold and Dershowitz (2018) \doi{10.1017/9781107415058}.
}
\author{
\strong{Maintainer}: Rob Hyndman \email{Rob.Hyndman@monash.edu} (\href{https://orcid.org/0000-0002-2140-5352}{ORCID}) [copyright holder]

}
\keyword{internal}
