% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinese.R
\name{chinese_date}
\alias{chinese_date}
\alias{as_chinese}
\title{Chinese dates}
\usage{
chinese_date(
  cycle = integer(),
  year = integer(),
  month = integer(),
  leap_month = logical(),
  day = integer()
)

as_chinese(date)
}
\arguments{
\item{cycle}{A numeric vector of cycles}

\item{year}{A numeric vector of years}

\item{month}{A numeric vector of months}

\item{leap_month}{A logical vector indicating leap months}

\item{day}{A numeric vector of days}

\item{date}{A numeric vector of dates}
}
\value{
A chinese vector object
}
\description{
Chinese dates
}
