% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinese.R
\name{chinese_new_year}
\alias{chinese_new_year}
\alias{dragon_festival}
\alias{qing_ming}
\title{Chinese holidays}
\usage{
chinese_new_year(year)

dragon_festival(year)

qing_ming(year)
}
\arguments{
\item{year}{The year on the Gregorian calendar}
}
\value{
A vector of dates on the Gregorian calendar
}
\description{
Dates are returned as Gregorian dates
}
\examples{
tibble::tibble(
  year = 2025:2030,
  cny = chinese_new_year(year),
  qm = qing_ming(year),
  dbf = dragon_festival(year)
)
}
\seealso{
\link{chinese_date}
}
