% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/christian.R
\name{christmas}
\alias{christmas}
\alias{advent}
\alias{epiphany}
\alias{eastern_orthodox_christmas}
\alias{orthodox_easter}
\alias{alt_orthodox_easter}
\alias{easter}
\alias{pentecost}
\title{Christian Ecclesiastical Holidays}
\usage{
christmas(year)

advent(year)

epiphany(year)

eastern_orthodox_christmas(year)

orthodox_easter(year)

alt_orthodox_easter(year)

easter(year)

pentecost(year)
}
\arguments{
\item{year}{Gregorian year}
}
\description{
Functions to return Gregorian dates for various Christian ecclesiastical holidays and other special days
}
