% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/christian.R
\name{advent}
\alias{advent}
\alias{christmas}
\alias{orthodox_christmas}
\alias{epiphany}
\alias{easter}
\alias{orthodox_easter}
\alias{alt_orthodox_easter}
\alias{pentecost}
\title{Christian Ecclesiastical Holidays}
\usage{
advent(year)

christmas(year)

orthodox_christmas(year)

epiphany(year)

easter(year)

orthodox_easter(year)

alt_orthodox_easter(year)

pentecost(year)
}
\arguments{
\item{year}{Gregorian year}
}
\description{
Functions to return Gregorian dates for various Christian ecclesiastical holidays and other special days
}
\examples{
tibble::tibble(
  year = 2025:2030,
  advent = advent(year),
  christmas = christmas(year),
  orthodox_christmas = orthodox_christmas(year),
  epiphany = epiphany(year),
  easter = easter(year),
  astronomical_easter = astronomical_easter(year),
  orthodox_easter = orthodox_easter(year),
  alt_orthodox_easter = alt_orthodox_easter(year),
  pentecost = pentecost(year)
)
}
