% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/christian.R
\name{advent}
\alias{advent}
\alias{christmas}
\alias{eastern_orthodox_christmas}
\alias{epiphany}
\alias{easter}
\alias{orthodox_easter}
\alias{alt_orthodox_easter}
\alias{pentecost}
\title{Christian Ecclesiastical Holidays}
\usage{
advent(year)

christmas(year)

eastern_orthodox_christmas(year)

epiphany(year)

easter(year)

orthodox_easter(year)

alt_orthodox_easter(year)

pentecost(year)
}
\arguments{
\item{year}{Gregorian year}
}
\description{
Functions to return Gregorian dates for various Christian ecclesiastical holidays and other special days
}
\examples{
advent(2025)
christmas(2025)
eastern_orthodox_christmas(2025)
epiphany(2025)
easter(2025)
orthodox_easter(2025)
alt_orthodox_easter(2025)
pentecost(2025)
}
