% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modern_hindu.R
\name{mesha_samkranti}
\alias{mesha_samkranti}
\alias{diwali}
\alias{shiva}
\alias{rama}
\alias{sacred_wednesdays}
\title{Hindu holidays and special days}
\usage{
mesha_samkranti(year)

diwali(year)

shiva(year)

rama(year)

sacred_wednesdays(year)
}
\arguments{
\item{year}{A numeric vector of Gregorian years}
}
\value{
A list of dates on the Gregorian calendar
}
\description{
Functions to return Gregorian dates for various Hindu holidays baed on the
Hindu lunar calendar. Diwali is on the new moon day in the month of Kartik (month 8).
The Great Night of Shiva is at the end of the 11 month of Magha.
The Birthday of Rama is on the 8th or 9th day of the first month (Caitra)
}
\examples{
shiva(2025)
rama(2025)
mesha_samkranti(2025)
diwali(2025)
sacred_wednesdays(2025)
}
\seealso{
\link{hindu_lunar_date}
}
