% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egyptian.R
\name{egyptian_date}
\alias{egyptian_date}
\alias{armenian_date}
\alias{as_egyptian}
\alias{as_armenian}
\title{Egyptian and Armenian dates}
\usage{
egyptian_date(year, month, day)

armenian_date(year, month, day)

as_egyptian(date)

as_armenian(date)
}
\arguments{
\item{year}{Numeric vector of years}

\item{month}{Numeric vector of months}

\item{day}{Numeric vector of days}

\item{date}{Vector of dates on some calendar}
}
\description{
Egyptian and Armenian dates
}
