% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granularities.R
\name{granularities}
\alias{granularities}
\alias{granularity}
\title{Find and extract canonical granularities from a calendar object}
\usage{
granularities(dates)

granularity(dates, granularity)
}
\arguments{
\item{dates}{A vector of calendar dates}

\item{granularity}{A character string indicating the granularity to extract}
}
\description{
\code{granularities()} will return a character vector of granularity names.
\code{granularity()} will return a vector of numerical values for a given granularity.
}
\examples{
granularities(iso_date(2025, 23, 2:4))
granularities(gregorian_date(2025, 4, 19:30))
granularities(as_roman(Sys.Date()))
granularity(iso_date(2025, 23, 2:4), "week")
granularity(gregorian_date(2025, 4, 19:21), "month")
granularity(as_roman(Sys.Date()), "event")
}
