% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granularities.R
\name{granularities}
\alias{granularities}
\alias{granularity}
\title{Canonical granularities}
\usage{
granularities(dates)

granularity(dates, granularity)
}
\arguments{
\item{dates}{A vector of calendar dates}

\item{granularity}{A character string indicating the granularity to extract}
}
\description{
\code{granularities()} will return a character vector of canonical granularity names for the
relevant calendar. These are the granularities used to define dates on the calendar.
\code{granularity()} will return a vector of numerical values for a given canonical granularity.
}
\examples{
granularities(iso_date(2025, 23, 2))
granularities(gregorian_date(2025, 4, 19))
granularities(as_roman(Sys.Date()))
granularity(iso_date(2025, 23, 2), "week")
granularity(gregorian_date(2025, 4, 19), "month")
granularity(as_roman(Sys.Date()), "count")
}
\seealso{
\code{\link{week_of_year}} for some non-canonical granularities.
}
