% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregorian.R
\docType{data}
\name{cal_gregorian}
\alias{cal_gregorian}
\alias{gregorian_date}
\alias{as_gregorian}
\title{Work with Gregorian calendar dates}
\format{
An object of class \code{calcal} of length 8.
}
\usage{
cal_gregorian

gregorian_date(year, month, day)

as_gregorian(date)
}
\arguments{
\item{year}{A numeric vector of years}

\item{month}{A numeric vector of months}

\item{day}{A numeric vector of days}

\item{date}{A gregorian vector object}
}
\value{
A gregorian vector object
}
\description{
Work with Gregorian calendar dates
}
\examples{
new_date(year = 2025, month = 3, day = 2:4, calendar = cal_gregorian)
as_date(Sys.Date(), calendar = cal_gregorian)
tibble::tibble(
  x = seq(as.Date("2025-01-01"), as.Date("2025-12-31"), by = "day"),
  z = as_date(x, calendar = cal_gregorian)
)
gregorian_date(2025, 4, 19:30)
as_gregorian("2016-01-01")
as_gregorian(Sys.Date())
tibble::tibble(
  x = seq(as.Date("2025-01-01"), as.Date("2025-12-31"), by = "day"),
  y = as_gregorian(x)
)
}
\keyword{datasets}
