% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gregorian.R
\name{gregorian_date}
\alias{gregorian_date}
\title{Gregorian calendar dates}
\usage{
gregorian_date(year = integer(), month = integer(), day = integer())
}
\arguments{
\item{year}{A numeric vector of years}

\item{month}{A numeric vector of months}

\item{day}{A numeric vector of days}
}
\value{
A gregorian vector object
}
\description{
Create a Gregorian date object.
}
\examples{
gregorian_date(2025, 4, 19:30)
}
