% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hebrew.R
\name{hebrew_date}
\alias{hebrew_date}
\alias{as_hebrew}
\title{Hebrew calendar dates}
\usage{
hebrew_date(year = integer(), month = integer(), day = integer())

as_hebrew(date)
}
\arguments{
\item{year}{A numeric vector of years}

\item{month}{A numeric vector of months}

\item{day}{A numeric vector of days}

\item{date}{Vector of dates on some calendar}
}
\value{
A hebrew vector object
}
\description{
Hebrew calendar dates
}
\examples{
heb <- new_date(year = 5785, month = 3, day = 2:4, calendar = cal_hebrew)
heb
hebrew_date(5785, 3, 2:4)
as_date(heb, calendar = cal_gregorian)
as_date(Sys.Date(), calendar = cal_hebrew)
tibble::tibble(
  gregorian = gregorian_date(2025, 1, 1) + 0:364,
  hebrew = as_date(gregorian, calendar = cal_hebrew),
)
as_gregorian(hebrew_date(5785, 3, 2:4))
as_hebrew(gregorian_date(2025, 1, 1:31))
as_hebrew("2016-01-01")
as_hebrew(Sys.Date())
}
\seealso{
\link{cal_hebrew}
}
