% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hebrew.R
\name{hebrew_date}
\alias{hebrew_date}
\alias{as_hebrew}
\title{Hebrew calendar dates}
\usage{
hebrew_date(year = integer(), month = integer(), day = integer())

as_hebrew(date)
}
\arguments{
\item{year}{A numeric vector of years}

\item{month}{A numeric vector of months}

\item{day}{A numeric vector of days}

\item{date}{Vector of dates on some calendar}

\item{...}{Additional arguments not currently used}
}
\value{
A hebrew vector object
}
\description{
Create a Hebrew date object.
}
\examples{
hebrew_date(2025, 3, 2:4)
as_gregorian(hebrew_date(2025, 3, 2:4))
as_hebrew(gregorian_date(2025, 1, 1:31))
as_hebrew("2016-01-01")
as_hebrew(Sys.Date())
tibble::tibble(
  x = seq(as.Date("2025-01-01"), as.Date("2025-12-31"), by = "day"),
  y = as_gregorian(x),
  z = as_hebrew(x)
)
}
