% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islamic.R
\docType{data}
\name{cal_islamic}
\alias{cal_islamic}
\alias{islamic_date}
\alias{as_islamic}
\title{Work with Islamic dates}
\format{
An object of class \code{calcal} of length 8.
}
\usage{
cal_islamic

islamic_date(year = integer(), month = integer(), day = integer())

as_islamic(date)
}
\arguments{
\item{year}{A numeric vector of years}

\item{month}{A numeric vector of months}

\item{day}{A numeric vector of days}

\item{date}{Vector of dates on some calendar}
}
\value{
An islamic vector object
}
\description{
Work with Islamic dates
}
\examples{
as_date("2016-01-01", calendar = cal_islamic)
as_date(Sys.Date(), calendar = cal_islamic)
tibble::tibble(
  x = new_date(year = 2025, month = 5, day = 1:31, calendar = cal_gregorian),
  y = as_date(x, calendar = cal_islamic)
)
new_date(year = 2025, month = 4, day = 19:30, calendar = cal_islamic)
islamic_date(2025, 4, 19:30)
as_islamic("2016-01-01")
as_islamic(Sys.Date())
tibble::tibble(
  x = seq(as.Date("2025-01-01"), as.Date("2025-12-31"), by = "day"),
  y = as_islamic(x)
)
}
\keyword{datasets}
