% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hebrew.R
\name{yom_kippur}
\alias{yom_kippur}
\alias{passover}
\alias{purim}
\alias{hanukkah}
\alias{rosh_hashanah}
\alias{sukkot}
\alias{shavuot}
\title{Jewish Holidays}
\usage{
yom_kippur(year)

passover(year)

purim(year)

hanukkah(year)

rosh_hashanah(year)

sukkot(year)

shavuot(year)
}
\arguments{
\item{year}{A numeric vector of Gregorian years}
}
\description{
Functions to return Gregorian dates for various Jewish holidays
}
\examples{
tibble::tibble(
  year = 2025:2030,
  rosh_hashanah = rosh_hashanah(year),
  passover = passover(year),
  shavuot = shavuot(year),
  yom_kippur = yom_kippur(year),
  sukkot = sukkot(year),
  hanukkah = hanukkah(year),
  purim = purim(year),
)
}
