% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bali.R
\name{kajeng_keliwon}
\alias{kajeng_keliwon}
\alias{tumpek}
\title{Balinese special days}
\usage{
kajeng_keliwon(year)

tumpek(year)
}
\arguments{
\item{year}{A numeric vector of Gregorian years}
}
\description{
Find all occurrences in a Gregorian year.
}
\seealso{
\link{balinese_date}
}
