% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_granularities.R
\name{linear_granularity}
\alias{linear_granularity}
\title{Linear Granularity}
\usage{
linear_granularity(gran_fun, n_gran, name)
}
\arguments{
\item{gran_fun}{A function that extracts a cyclic granularity from a date
variable. This function must return an integer or an ordered factor.}

\item{n_gran}{An integer giving the number of granularities per year.}

\item{name}{A character string giving the name of the granularity}
}
\value{
A function that creates a linear granularity object.
}
\description{
Linear granularities are produced by combining the year and a cyclic granularity.
}
\examples{
# Compute quarter from a Gregorian date
quarter <- function(date) {
  factor(ceiling(month_of_year(date) / 3),
    levels = 1:4,
    labels = c("Q1", "Q2", "Q3", "Q4"),
    ordered = TRUE
  )
}
year_quarter <- linear_granularity(quarter, 4, "year-quarter")
year_quarter(gregorian_date(2025, 6, 25) + 1:10)
}
