% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarlunar.R
\name{lunar_phase}
\alias{lunar_phase}
\title{Lunar phase at date}
\usage{
lunar_phase(date)
}
\arguments{
\item{date}{Date vector}
}
\description{
Lunar phase at date, as an angle in degrees. An angle of 0 means a new moon,
90 degrees means the first quarter, 180 means a full moon, and 270 degrees
means the last quarter.
}
\examples{
april2025 <- gregorian_date(2025,4,1:30)
lunar_phase(april2025)

}
