% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_hindu.R
\name{old_hindu_solar_date}
\alias{old_hindu_solar_date}
\alias{as_old_hindu_solar}
\title{Old Hindu Solar dates}
\usage{
old_hindu_solar_date(year, month, day)

as_old_hindu_solar(date)
}
\arguments{
\item{year}{A numeric vector of years}

\item{month}{A numeric vector of months}

\item{day}{A numeric vector of days}

\item{date}{A date vector on some calendar}
}
\description{
Old Hindu Solar dates
}
\examples{
gregorian_date(2025, 1, 1:31) |>
  as_old_hindu_solar()
}
