% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roman.R
\docType{data}
\name{cal_roman}
\alias{cal_roman}
\alias{roman_date}
\alias{as_roman}
\title{Work with Roman calendar dates}
\format{
An object of class \code{calcal} of length 8.
}
\usage{
cal_roman

roman_date(year, month, event, count, leap)

as_roman(date)
}
\arguments{
\item{year}{A numeric vector of years}

\item{month}{A numeric vector of months}

\item{event}{A numeric vector of events: 1 = KALENDS, 2 = NONES, 3 = IDES}

\item{count}{A numeric vector of counts:}

\item{leap}{A logical vector indicating if year is a leap year}

\item{date}{Vector of dates on some calendar}
}
\value{
A roman vector object
}
\description{
Work with Roman calendar dates
}
\examples{
roman_date(66, 4, 1, 1, FALSE)
new_date(year = 66, month = 4, event = 1, count = 1, leap = FALSE, calendar = cal_roman)
as_roman("2016-01-01")
tibble::tibble(
  x = seq(as.Date("2025-01-01"), as.Date("2025-12-31"), by = "day"),
  y = as_roman(x)
)
}
\keyword{datasets}
