% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astronomical.R
\name{sunrise}
\alias{sunrise}
\alias{sunset}
\alias{asr}
\title{Sunrise, sunset and asr given a date and location}
\usage{
sunrise(date, locale, ...)

sunset(date, locale, ...)

asr(date, locale, ...)
}
\arguments{
\item{date}{Date in rd_fixed format}

\item{locale}{Location of class "location", usually the output from the \code{location} function}

\item{...}{Additional arguments passed to specific methods}
}
\value{
Time of sunrise
}
\description{
Calculate the time of sunrise and sunset at a specific location and date
}
\examples{
melbourne <- location(-37.8136, 144.9631, 31, 10)
sydney <- location(-33.8688, 151.2093, 3, 10)
sunrise("2025-01-01", c(melbourne, sydney))
sunset("2025-01-01", c(melbourne, sydney))
asr("2025-01-01", c(melbourne, sydney))
}
