% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astronomical.R
\name{sunrise}
\alias{sunrise}
\title{Sunrise given a date and location}
\usage{
sunrise(date, locale)
}
\arguments{
\item{date}{Date in rd_fixed format}

\item{locale}{Location of class "location", usually the output from the \code{location} function}
}
\value{
Time of sunrise
}
\description{
Calculate the time of sunrise at a specific location and date
}
