% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astronomical.R
\name{sunrise}
\alias{sunrise}
\alias{sunset}
\title{Sunrise and sunset given a date and location}
\usage{
sunrise(date, location, ...)

sunset(date, location, ...)
}
\arguments{
\item{date}{Date in calcalvec format}

\item{location}{Location of class "location", usually the output from the \code{location} function}

\item{...}{Additional arguments passed to specific methods}
}
\value{
Time of sunrise
}
\description{
Calculate the time of sunrise and sunset at a specific location and date. The
time zone of the location is used as specified in the \code{location} object.
}
\examples{
melbourne <- location(-37.8136, 144.9631, 31, 10)
sydney <- location(-33.8688, 151.2093, 3, 10)
sunrise("2025-01-01", c(melbourne, sydney))
sunset("2025-01-01", c(melbourne, sydney))
}
