% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{as_time_of_day}
\alias{as_time_of_day}
\title{Convert to time of day}
\usage{
as_time_of_day(x, ...)
}
\arguments{
\item{x}{Vector of times}

\item{...}{Additional arguments not currently used}
}
\description{
Convert to time of day
}
\examples{
as_time_of_day(Sys.time())
}
